+++ satella/coding/transforms/jsonify.py
@@ -1,6 +1,7 @@
 import typing as tp
 
 from satella.json import JSONAble
+import enum
 
 
 def jsonify(data: tp.Any) -> tp.Optional[tp.Union[str, int, float, list, dict]]:
@@ -8,6 +9,8 @@ def jsonify(data: tp.Any) -> tp.Optional[tp.Union[str, int, float, list, dict]]:
     Convert any data to a value that's serializable via JSON.
 
     Objects that are JSONAble will have their to_json() method called.
+
+    Note that enums will be converted to their value.
 
     As a last resort, str() will be called on the object, and if that fails
     it will have repr() called on it
@@ -19,6 +22,8 @@ def jsonify(data: tp.Any) -> tp.Optional[tp.Union[str, int, float, list, dict]]:
         return None
     elif isinstance(data, (int, float, str)):
         return data
+    elif isinstance(data, enum.Enum):
+        return data.value
     elif isinstance(data, JSONAble):
         return jsonify(data.to_json())
     elif isinstance(data, tp.Mapping):