+++ satella/coding/decorators/__init__.py
@@ -3,8 +3,10 @@ from .arguments import auto_adapt_to_methods, attach_arguments, for_argument, \
 from .decorators import wraps, chain_functions, has_keys, short_none, memoize
 from .flow_control import loop_while, queue_get
 from .preconditions import postcondition, precondition
+from .retry_dec import retry
 
-__all__ = ['execute_before', 'postcondition', 'precondition', 'wraps', 'queue_get',
+__all__ = ['retry',
+           'execute_before', 'postcondition', 'precondition', 'wraps', 'queue_get',
            'chain_functions', 'has_keys', 'short_none', 'auto_adapt_to_methods',
            'attach_arguments', 'for_argument', 'loop_while', 'memoize',
            'copy_arguments', 'replace_argument_if']