+++ satella/coding/decorators/retry_dec.py
@@ -1,11 +1,12 @@
+import itertools
 import typing as tp
 
 from satella.coding.decorators.decorators import wraps
 from satella.coding.typing import ExceptionClassType
 
 
-def retry(times: int,
-          exc_classes: tp.Union[ExceptionClassType, tp.Tuple[ExceptionClassType, ...]],
+def retry(times: tp.Optional[int] = None,
+          exc_classes: tp.Union[ExceptionClassType, tp.Tuple[ExceptionClassType, ...]] = Exception,
           on_failure: tp.Callable[[Exception], None] = lambda e: None,
           swallow_exception: bool = True):
     """
@@ -26,8 +27,9 @@ def retry(times: int,
     >>>     except RuntimeError:
     >>>         pass
 
-    :param times: maximum times to retry this operation
-    :param exc_classes: exception classes that should be considered failure
+    :param times: maximum times to retry this operation. By default (None) retry until the world ends.
+    :param exc_classes: exception classes that should be considered failure. By default.
+        catches everything (Exception).
     :param on_failure: callable to call when it fails times times. Callable will be called
         with a single argument, exception instance that was raised last. That exception will
         be swallowed, unless swallow_exception is set to False
@@ -37,7 +39,11 @@ def retry(times: int,
     def outer(fun):
         @wraps(fun)
         def inner(*args, **kwargs):
-            for i in range(times):
+            if times is None:
+                iterator = itertools.count()
+            else:
+                iterator = range(times)
+            for _ in iterator:
                 try:
                     return fun(*args, **kwargs)
                 except exc_classes as e: