+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # v2.14.4
 
+* added exception `Empty`
+* satella futures are now generic
+* added `Closeable`
+* added `PeekableQueue`