+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.5
+
+* `Closeable` is now in __all__