+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * `Closeable` is now in __all__
 * `Closeable` has __slots__
+* fixed a bug in `PeekableQueue`
+* `TerminableThread.safe_sleep` will do the right thing when passed a negative value