+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.6
+
+* `CPUTimeAwareIntervalTerminableThread` max_sooner will now by default be 10% of seconds