+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.7
+
+* specified default arguments for retry