+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.14.8
 
+* **bugfix** `@wraps` would store not a `__dict__` in `__annotations__`