+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-# v2.14.8
+# v2.14.9
 
-* **bugfix** `@wraps` would store not a `__dict__` in `__annotations__`